Improved slicing to cope with units.
The slicing capabilities of the FieldContainer were improved to support
various notations of slicing with units. Basically for any axis a string can be
given that follows the python slice notation with a twist: Every number
can carry a unit, or the groups as a whole can carry a unit, in which case
it is simply attached to the last number.


git-svn-id: https://pyphant.svn.sourceforge.net/svnroot/pyphant/trunk@463 3321fe91-ea1f-0410-9b6b-81c2d13ccd49