Added alternative symbol for exponentiation.
For PhysicalQuantities a to the bth must be written
as a**b. However the notation a^b is also very common
and in our setting unlikely to conflict with the other meaning
of ^, name the exclusive or. Now a^b will be recognized as
pow(a,b) both by the parser and by the DataContainer.


git-svn-id: https://pyphant.svn.sourceforge.net/svnroot/pyphant/trunk@486 3321fe91-ea1f-0410-9b6b-81c2d13ccd49