Multiple dependency support in single .fmf files.
This adds support for multiple dependencies in .fmf files,
i.e. with three fields a, b, c, now c can depend on both
a and b and will be reshaped accordingly.
The field will be generated by unique(a) x unique(b) and
missing cells will be none. The mask will also be set accordingly.


git-svn-id: https://pyphant.svn.sourceforge.net/svnroot/pyphant/trunk@489 3321fe91-ea1f-0410-9b6b-81c2d13ccd49