viewOSC visualizes the thickness map
Setting parameter '-v thicknessMap' shows the thickness map of the
investigated sample. Furthermore the position of the absorption curve,
which is specified by parameter -n, is indicated by plotting a cross at the
respective position. For this purpose
the coordinate system of the thickness map is shifted by -(stepX,stepY),
such that the cross marks the center of the respective pixel.


git-svn-id: https://pyphant.svn.sourceforge.net/svnroot/pyphant/trunk@547 3321fe91-ea1f-0410-9b6b-81c2d13ccd49