Bugfix Skeletonize worker
The Skeletonize worker correctly handles curved and regular features
correctly by selecting all pixel for which rules I, II and III apply.
These pixels are checked one by one for rule IV and an additional fifth rule (a pixel can be removed if was not
located diagonally to a corner feature pixel in the previous iteration).
Pixel which do not pass this test in the first run are checked within a
second run against rules IV and V.


git-svn-id: https://pyphant.svn.sourceforge.net/svnroot/pyphant/trunk@586 3321fe91-ea1f-0410-9b6b-81c2d13ccd49