Enh: Started to implement tests for H5FileHandler
At this point there is only one test that expects an IOError
when a file that does not exist is opened in readonly mode.
Also fixed __del__() method in H5FileHandler to check whether
the handle attribute exists before closing it. This is necessary
if the file could not be opened as in the above case.