Enh: Saving and loading of DCs added in H5FileHandler
Also wrote tests for saving and loading of DataContainers.
TestH5FileHandler.py is clean but H5FileHandler.py is still a mess
and has been created by copy and paste from PyTablesPersister.py
Usage: Create one instance of a H5FileHandler class per HDF5 file and
handle all IO via that instance.
The constructor parameters are: H5FileHandler(filename, mode = 'a')
where filename is the path to the HDF5 file to be opened and mode
is in ['r', 'w', 'a'] meaning readonly, write and append. If mode
is set to 'r' or 'a' the file given by filename has to exist. In
the case of mode being set to 'a' the file has to be a valid HDF5
file already. 'w' is usally used when the file does not exist.
Then a new empty file is created and gets a valid HDF5 file as soon
as there is data written to it successfully. If the file already exist
and mode is set to 'w' its content is deleted first.
DataContainers that have been sealed or at least carry a valid emd5 can
be saved with the saveDataContainer(dc) method, where dc is a
DataContainer Instance. A HDF5 group path to the storage location is
returned by this method but can be discarded if all IO is handled via the
H5FileHandler (which is the recommended procedure).
DataContainers that are present in the HDF5 file can be loaded with the
loadDataContainer(dcId) method, where dcId is the emd5 of the DataContainer
to be loaded.
Both saving and loading of DataContainers works for Field- and
SampleContainers and has been tested successfully in TestH5FileHandler.py