Enh: Added loadSummary method to H5FileHandler
A summary including meta data of a given DataContainer can be loaded
from HDF5 files using the H5FileHandler. Usage is as follows:
loadSummary(dcId) returns a dictionary filled with meta data of the given
DataContainer. Here dcId is the emd5 of the DataContainer. The information
is extracted from the HDF5 file without loading the entire DataContainer
itself. The keys of the dictionary are: 'id', 'longname', 'shortname',
'attributes', the values are self explanatory. In case of FieldContainers
there are additional keys:
'unit', 'dimensions', where the value of 'dimensions' is a list of
summary dictionaries of the dimensions of the FieldContainer. If there
are no dimensions (e.g. if the FieldContainer itself was a dimension),
'dimension' is set to u'INDEX'.
For SampleContainers, there is one additional key 'columns' and its
value is a list of summary dictionaries of the columns.
If dcId was set to None (which is the default value), a dictionary
mapping emd5s to summary dictionaries of all available DataContainers
in the HDF5 file is returned.
A test for the loadSummary method has been added to TestH5FileHandler.py
and it passes successfully.