Cosm: Separated WebInterface into new module.
There is a new module called 'WebInterface' that provides the
WebInterface class and some additional helper classes that were
located in the KnowledgeManager module before. This module handles
the internal HTTP and HTML communication of the KnowledgeManager and
also is intended to contain the Pyphant web frontend which has been
disabled by default and its content has been removed due to changes in
getSummary(). It can be enabled by setting
km.web_interface.disabled = False
where km is a KnowledgeManager instance. But as said before there is no
functionality at all at this point.
The part for handling HTTP POST and GET requests is still located in
the KnowledgeManager module in order to avoid circular imports since
that part essentially depends on the KnowledgeManager.