Enh: Implemented bounding box matching in AutoFocus worker
This adds bounding box overlap matching for the AutoFocus worker.
The implementation was done using a Cube class that provides 'and' and
'or' operations for n-dimensional cubes that are build of n slices.
Tests are going to be written soon. This version is not considered to be
working at all.
Matching by feature overlap is still to come also.