Enh: Added tests for AutoFocus.py
These tests cover creation, intersection, concatenation and subtraction of
cubes as well as volume and edge length measurement.
For the zTube class, threre is a test for bounding box matching of features.
It is also tested, whether matching features are sorted into the zTube
properly.
Implemented tests for the Cube and zTube class