Fix: Changed source of z-values for AutoFocus
The z-values now arise from the input FieldContainer of the AutoFocus worker.
This FieldContainer should consist of a one dimensional FC with z-values along
the 0st axis. Each value of the FC should be another FC as produced by the
output of the MeasureFocus worker (with parameter 'HumanOutput' disabled).
The input FC of the AutoFocus worker should also carry an attribute with
key u'ztol' and a PhysicalQuantity as value that corresponds to the thickness
of the z slices. This value should be set so that
slice(zvalue1 - ztol, zvalue1 + ztol) and slice(zvalue2 - ztol, zvalue2 + ztol)
have nonempty intersection if and if only zvalue1 and zvalue2 correspond to
adjacent layers. It is also possible that ztol is estimated automatically from
the 0st axis of the input FC in a future version.