Enh: Changed attributes of FocusSlices
The attribute 'label' has been removed since it is not needed.
The attribute 'area' has been replaced by 'mask' in order to allow more
general methods of determining a mid point and radius. The mask attribute
carries a numpy.array of dtype bool and shape according to the
slice attribute.