Fix: Fixed input requirement + bugs in AutoFocus worker
Since storing FieldContainers within FieldContainers is not a good idea at all,
the input requirement has changed. The AutoFocus worker now expects a
SampleContainer with two columns as input. The first column bares the z values
and the second column bares the corresponding emd5s of the results of the
MeasureFocus worker.
Also fixed some bugs concerning use of PhysicalQuantities.