Enh: Started implementing BatchWorker
There is a new worker in the tools toolbox that is capable of calculating
a sequence of results from a certain plug in the recipe for different input
and/or different parameters.
(Parameter mapping is not yet implemented.)
The worker expects a SC as input that has a column 'emd5'.
It calculates the result for each input DC in the emd5 column and creates
a copy of the input SC as output, with the emd5s replaced by the emd5s of the
results.
Output worker and plug can be selected by parameter dropdown.
It is necessary to press 'OK' after each selection to force a call to
refreshParams(). For example: Select an output worker. Press 'OK'.
In the next step you can select the output plug from all the plugs that worker
provides.
The input socket is determined as the zerost open socket that leads to the
selected output plug. More options to select from several open sockets will
come in future implementations.
There is also parameter mapping planned:
You can select how to map parameters:
Either don't change parameters, just iterate over different input.
Or map a column from the input SC to a Parameter of the input worker.
Or map all columns (except 'emd5') of the input SC to the according parameters
of the input worker. (longname of the colum has to match the input worker's
parameter name)
Parameter mapping is selectable in the BatchWorker but not yet implemented,
as said before.
There is another new Worker planned that allows to extract DataContainers
from the BatchWorker's output. (This is necessary for better human handling
since there are only emd5s stored in the output.)