Enh: Added BatchExtractor worker
This worker allows to extract DataContainers from the resulting SampleContainer
of the BatchWorker.
The only parameter is 'Index' which is set to zero by default and determines
the row in the input SampleContainer from where to extract the actual data,
i.e. resolve the entry in the 'emd5' column to a actual DataContainer.
Possible use:
Plug 'SC Filter' worker to the output of a BatchWorker to select your
favourite data e.g. '"myParameter" == 4.3 nm', plug the result of 'SC Filter'
to a BatchExtractor (Index set to 0) and you get your data.