Enh: Added ParameterRun worker
Usage:
Plug a data source into the socket of the worker.
Connect the plug of the worker to a BatchWorker.
Adjust start, stop, step, unit and set the parameter name.
Select "source/dest" as mapping method in the BatchWorker.
Select your chosen parameter name as source.
Adjust the BatchWorker as usual and set the parameter you want to control
as destination.
Use a BatchExtractor worker (best in combination with SC filter worker)
to visualize the individual results of the run.