Enh: More options for parameter mapping in BatchWorker
The parameter destination is not restricted to the input worker anymore.
Now there is an option to select any parameter of any worker in the recipe
to remote control with the BatchWorker.
There is also a new switch that allows to convert parameters to strings.
(Useful for parameters with units as they usually are stored as user input
strings)