Enh: Replaced BatchWorker by BatchHead and BatchTail
There are two new workers that replace the BatchWorker.
Usage:
Put the BatchHead worker between an adequate data source (e.g. a
ParameterRun worker) and the entrypoint to your recipe where you would
usually put a single data source.
Configure the BatchHead worker to map parameters and so on...
Connect a BatchTail worker anywhere to your recipe where you want to
get your result from and optionally connect a BatchExtractor to it in
order to view a single result of the sequence.
Click on the BatchTail worker and select the proper BatchHead.
Once a BatchHead has been selected, every click to the BatchTail worker
forces the calculation of all the results.
You can then visualize them as usual using the BatchExtractor.