Enh: Added switch for WebInterface to wxPyphant
There is a switch in the file menu that toggles the WebInterface (and
KM http server). The WebInterface and KM server are still disabled by
default. The server listens on 127.0.0.1:8000 once it has been enabled.
In addition, the frontpage of the WebInterface pops up in a browser
when the server is started.