Fix: Fixed 'too many open files' bug
The H5FileHandler now provides __enter__ and __exit__ methods and must be
used with the 'with' statement from now on.
This ensures that HDF files are opened before an operation and closed after
the operation or if an error occurs.
This fixes the bug of too many HDF files being open at one time.
Usage example:
handler = H5FileHandler('dummy.h5', 'r')
with handler:
    dc = handler.loadDataContainer(emd5)