Enh: Wrote XMLHandler
There is a method getXMLRoot(filename_or_stream) in the
XMLHandler module that returns an XMLElement instance that
resembles the root element in the given XML file.
An XMLElement object provides the following attributes:
- name
- attributes
- content
- parent
- children
That is all XML elements in the file can be accessed via
the root element. '__getitem__' and '__setitem__' methods
are implemented for XMLElement to access children.