Enh: ImageProcessing Reduce Worker
Implemented a Reduce Worker in the ImageProcessing
toolbox. The worker takes a stack of images
and reduces it to a single image using a binary
operation on images. At this time, the operations
or, and and xor are implemented, but the worker
can easily be extended to support any other
operation that has a calcStep method that
accepts two FieldContainers and returns a single
FieldContainer.