Enh: Added Algebra worker, ExtractRows worker
The Algebra worker allows to perform calculations involving columns
of a SampleContainer. The functionality of the old filter worker is
regained by a combination of the Algebra worker and the ExtractRows
worker by using an expression that evaluates to a boolean.

The ExtractRows worker takes a SampleContainer and a FieldContainer
(the mask) as input and creates a new SampleContainer containing only
the rows where mask is set to True.

See the docstrings of the methods
calcColumn, filter, extractRows of the SampleContainer class for
a detailed description of possible expressions.

Some basic tests have been implemented.

Signed-off-by: Alexander Held <alexander.held@fmf.uni-freiburg.de>